+++ BUILD.bazel
@@ -10,7 +10,9 @@ genrule(
         "//c-toxcore/toxcore:tox.h",
         "//c-toxcore/toxcore:tox_dispatch.h",
         "//c-toxcore/toxcore:tox_events.h",
+        "//c-toxcore/toxcore:tox_options.h",
         "//c-toxcore/toxcore:tox_private.h",
+        "//c-toxcore/toxcore:tox_types.h",
         "//c-toxcore/toxencryptsave:toxencryptsave.h",
     ],
     outs = [
@@ -18,7 +20,9 @@ genrule(
         "tox/tox.h",
         "tox/tox_dispatch.h",
         "tox/tox_events.h",
+        "tox/tox_options.h",
         "tox/tox_private.h",
+        "tox/tox_types.h",
         "tox/toxencryptsave.h",
     ],
     cmd = """
@@ -26,7 +30,9 @@ genrule(
         cp $(location //c-toxcore/toxcore:tox.h) $(GENDIR)/c-toxcore/tox/tox.h
         cp $(location //c-toxcore/toxcore:tox_dispatch.h) $(GENDIR)/c-toxcore/tox/tox_dispatch.h
         cp $(location //c-toxcore/toxcore:tox_events.h) $(GENDIR)/c-toxcore/tox/tox_events.h
+        cp $(location //c-toxcore/toxcore:tox_options.h) $(GENDIR)/c-toxcore/tox/tox_options.h
         cp $(location //c-toxcore/toxcore:tox_private.h) $(GENDIR)/c-toxcore/tox/tox_private.h
+        cp $(location //c-toxcore/toxcore:tox_types.h) $(GENDIR)/c-toxcore/tox/tox_types.h
         cp $(location //c-toxcore/toxencryptsave:toxencryptsave.h) $(GENDIR)/c-toxcore/tox/toxencryptsave.h
     """,
     visibility = ["//visibility:public"],
