+++ cmake/StrictAbi.cmake
@@ -32,6 +32,9 @@ function(_make_version_script target)
       COMMAND ${SHELL} -c "egrep '^\\w' ${header} | grep '${ns}_[a-z0-9_]*(' | grep -v '^typedef' | grep -o '${ns}_[a-z0-9_]*(' | egrep -o '[a-z0-9_]+' | sort -u"
       OUTPUT_VARIABLE sublib_SYMS
       OUTPUT_STRIP_TRAILING_WHITESPACE)
+    if("${sublib_SYMS}" STREQUAL "")
+      continue()
+    endif()
     string(REPLACE "\n" ";" sublib_SYMS ${sublib_SYMS})
 
     foreach(sym ${sublib_SYMS})
