+++ other/docker/modules/check
@@ -12,24 +12,24 @@ MODS = {}
 STD_MODULE = """module std [system] {
   textual header "/usr/include/alloca.h"
   textual header "/usr/include/assert.h"
-  textual header "/usr/include/c++/13.2.1/algorithm"
-  textual header "/usr/include/c++/13.2.1/array"
-  textual header "/usr/include/c++/13.2.1/chrono"
-  textual header "/usr/include/c++/13.2.1/cstddef"
-  textual header "/usr/include/c++/13.2.1/cstdint"
-  textual header "/usr/include/c++/13.2.1/cstdio"
-  textual header "/usr/include/c++/13.2.1/cstdlib"
-  textual header "/usr/include/c++/13.2.1/cstring"
-  textual header "/usr/include/c++/13.2.1/iomanip"
-  textual header "/usr/include/c++/13.2.1/iosfwd"
-  textual header "/usr/include/c++/13.2.1/limits"
-  textual header "/usr/include/c++/13.2.1/memory"
-  textual header "/usr/include/c++/13.2.1/ostream"
-  textual header "/usr/include/c++/13.2.1/random"
-  textual header "/usr/include/c++/13.2.1/stdlib.h"
-  textual header "/usr/include/c++/13.2.1/thread"
-  textual header "/usr/include/c++/13.2.1/type_traits"
-  textual header "/usr/include/c++/13.2.1/vector"
+  textual header "/usr/include/c++/14.2.0/algorithm"
+  textual header "/usr/include/c++/14.2.0/array"
+  textual header "/usr/include/c++/14.2.0/chrono"
+  textual header "/usr/include/c++/14.2.0/cstddef"
+  textual header "/usr/include/c++/14.2.0/cstdint"
+  textual header "/usr/include/c++/14.2.0/cstdio"
+  textual header "/usr/include/c++/14.2.0/cstdlib"
+  textual header "/usr/include/c++/14.2.0/cstring"
+  textual header "/usr/include/c++/14.2.0/iomanip"
+  textual header "/usr/include/c++/14.2.0/iosfwd"
+  textual header "/usr/include/c++/14.2.0/limits"
+  textual header "/usr/include/c++/14.2.0/memory"
+  textual header "/usr/include/c++/14.2.0/ostream"
+  textual header "/usr/include/c++/14.2.0/random"
+  textual header "/usr/include/c++/14.2.0/stdlib.h"
+  textual header "/usr/include/c++/14.2.0/thread"
+  textual header "/usr/include/c++/14.2.0/type_traits"
+  textual header "/usr/include/c++/14.2.0/vector"
   textual header "/usr/include/errno.h"
   textual header "/usr/include/fortify/stdio.h"
   textual header "/usr/include/fortify/string.h"
