+++ other/docker/modules/run
@@ -1,6 +1,3 @@
-#!/bin/sh
+#!/usr/bin/env bash
 
-set -eux
-BUILD=modules
-other/docker/sources/build
-docker build -t "toxchat/c-toxcore:$BUILD" -f "other/docker/$BUILD/$BUILD.Dockerfile" .
+. "$(cd "$(dirname "${BASH_SOURCE[0]}")/../sources" && pwd)/run.sh"
