+++ other/docker/pkgsrc/dockerignore
@@ -1,0 +1,1 @@
+!other/docker/pkgsrc/pkgsrc.patch
