+++ other/docker/pkgsrc/pkgsrc.Dockerfile
@@ -1,11 +1,14 @@
 FROM toxchat/pkgsrc:latest
 
 WORKDIR /work
-COPY . /work/c-toxcore-0.2.18
-RUN ["tar", "zcf", "c-toxcore.tar.gz", "c-toxcore-0.2.18"]
+COPY . /work/c-toxcore-0.2.20
+RUN ["tar", "zcf", "c-toxcore.tar.gz", "c-toxcore-0.2.20"]
+
+WORKDIR /work/pkgsrc
+COPY other/docker/pkgsrc/pkgsrc.patch /tmp/pkgsrc.patch
+RUN ["patch", "-p1", "-i", "/tmp/pkgsrc.patch"]
 
 WORKDIR /work/pkgsrc/chat/toxcore
-RUN ["sed", "-i", "-e", "s/libtoxcore.so.2.18.0/libtoxcore.so.2.20.0/g", "PLIST"]
 RUN ["bmake", "clean"]
 RUN ["bmake", "DISTFILES=c-toxcore.tar.gz", "DISTDIR=/work", "NO_CHECKSUM=yes"]
 RUN ["bmake", "install"]
