+++ other/docker/pkgsrc/pkgsrc.patch
@@ -1,0 +1,31 @@
+diff --git a/chat/toxcore/Makefile b/chat/toxcore/Makefile
+index 70466704d..53a08ad08 100644
+--- a/chat/toxcore/Makefile
++++ b/chat/toxcore/Makefile
+@@ -1,6 +1,6 @@
+ # $NetBSD: Makefile,v 1.6 2024/01/22 13:16:56 ryoon Exp $
+
+-DISTNAME=	toxcore-0.2.18
++DISTNAME=	toxcore-0.2.20
+ PKGREVISION=	2
+ CATEGORIES=	chat
+ MASTER_SITES=	${MASTER_SITE_GITHUB:=TokTok/}
+diff --git a/chat/toxcore/PLIST b/chat/toxcore/PLIST
+index f0a5e4f04..4122b0867 100644
+--- a/chat/toxcore/PLIST
++++ b/chat/toxcore/PLIST
+@@ -4,11 +4,13 @@ bin/tox-bootstrapd
+ include/tox/tox.h
+ include/tox/tox_dispatch.h
+ include/tox/tox_events.h
++include/tox/tox_options.h
++include/tox/tox_types.h
+ include/tox/toxav.h
+ include/tox/toxencryptsave.h
+ lib/libtoxcore.a
+ lib/libtoxcore.so
+ lib/libtoxcore.so.2
+-lib/libtoxcore.so.2.18.0
++lib/libtoxcore.so.2.20.0
+ lib/pkgconfig/toxcore.pc
+ share/bash-completion/completions/tox-bootstrapd
