+++ toxcore/BUILD.bazel
@@ -6,7 +6,9 @@ exports_files(
         "tox.h",
         "tox_dispatch.h",
         "tox_events.h",
+        "tox_options.h",
         "tox_private.h",
+        "tox_types.h",
     ],
     visibility = ["//c-toxcore:__subpackages__"],
 )
@@ -1073,6 +1075,25 @@ cc_library(
 )
 
 cc_library(
+    name = "tox_types",
+    srcs = ["tox_types.c"],
+    hdrs = ["tox_types.h"],
+    visibility = ["//c-toxcore:__subpackages__"],
+)
+
+cc_library(
+    name = "tox_options",
+    srcs = ["tox_options.c"],
+    hdrs = ["tox_options.h"],
+    copts = ["-UTOX_HIDE_DEPRECATED"],
+    visibility = ["//c-toxcore:__subpackages__"],
+    deps = [
+        ":ccompat",
+        ":tox_types",
+    ],
+)
+
+cc_library(
     name = "tox",
     srcs = [
         "tox.c",
@@ -1104,6 +1125,8 @@ cc_library(
         ":network",
         ":onion_client",
         ":state",
+        ":tox_options",
+        ":tox_types",
         ":util",
         "//c-toxcore/toxencryptsave:defines",
         "@pthread",
@@ -1117,6 +1140,7 @@ cc_test(
     deps = [
         ":crypto_core",
         ":tox",
+        ":tox_options",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
