+++ toxcore/tox_test.cc
@@ -6,6 +6,7 @@
 #include <vector>
 
 #include "crypto_core.h"
+#include "tox_options.h"
 #include "tox_private.h"
 
 namespace {
