+++ toxcore/tox_types.c
@@ -1,0 +1,6 @@
+/* SPDX-License-Identifier: GPL-3.0-or-later
+ * Copyright © 2016-2025 The TokTok team.
+ * Copyright © 2013 Tox project.
+ */
+
+#include "tox_types.h" // IWYU pragma: keep
